requirements.txt
alembic==0.7.6
blinker==1.4
dominate==2.1.12
Flask==0.10.1
Flask-Bootstrap==3.3.5.3
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
python-chess==0.9.1
SQLAlchemy==1.0.6
Werkzeug==0.10.4
WTForms==2.0.2
